Move R_getVar() and R_getVarEx() to utils.h
