+++ src/cleancall.c
@@ -2,6 +2,7 @@
 #include <Rinternals.h>
 
 #include "cleancall.h"
+#include "utils.h"
 
 
 #if (defined(R_VERSION) && R_VERSION < R_Version(3, 4, 0))
@@ -14,23 +15,6 @@
    fn_ptr ptr;
    ptr.p = R_ExternalPtrAddr(s);
    return ptr.fn;
- }
-#endif
-
-#if (defined(R_VERSION) && R_VERSION < R_Version(4, 5, 0))
- SEXP R_getVar(SEXP symbol, SEXP rho, Rboolean inherits) {
-   SEXP out;
-   if (inherits) {
-     out = Rf_findVar(symbol, rho);
-   } else {
-     out = Rf_findVarInFrame(rho, symbol);
-   }
-
-   if (out == R_UnboundValue) {
-     const char *name = CHAR(PRINTNAME(symbol));
-     Rf_error("object '%s' not found", name);
-   }
-   return out;
  }
 #endif
